FK, FABRIK and CCD
finished the Forward kinematics section of the project, the armature can be moved by using the arrow keys, '+' and '-' keys are used for going up or down the armature bones.

CCD work halfway done, the armature moves, but it seems to ignore directions of wether to rotate in a positive or negative angle.

can swap between FABRIK and CCD using the 'c' key