FABRIK fix, screen to world mouse input, spring fix
fixed a bit of the fabrik algorith, fixed a large part of the CCD algorithm. improved the spring by adding gravity together with velocity members and changing other parameters.

linked the spring to the lt bone of the armature.

To use FK, use the arrow keys to move left or right, use '-' and '+' to move up and down the hierarchy.

To use IK, click and drag on the screen, use 'C' to swap between FABRIK and CCD.

the spring anchor is the black node, the weight tends to move out of the screen when being instantiated, give it some time to return to its resting place