Scene fix and first camera fix
separated the scene from the scene manager, allowing for multiple scenes to exist at the same time and have one be active at a time.

added a fix t allow for camera rotation using the view matrix. similar implementation for camera movement added to the camera, however it does not currently work