Shader Manager Fix & new CBuffer bug?
the shader manager has been fixed, however, there seems to be a problem when updating constant buffers.

my main reason for thinking this is the issue is because of 2 passes: the normal view pass and the emissive blur pas, in the case of the normal view pass, there seems to be an issue when updating the view matrix, that prevents positions from being stored, and in the emissive pass, it uses the exact same shaders as the luminance pass, however, while blurring luminance works without issues, the emissive pass does not do anything.